+++ pipelines/prosmart_refmac/script/prosmart_refmac.py
@@ -25,6 +25,9 @@ from core.CCP4PluginScript import CPluginScript
 from core import CCP4ErrorHandling
 from core import CCP4Utils
 import os,sys,shutil,re
+import traceback
+from wrappers.modelASUCheck.script.modelASUCheck import sequenceAlignment
+
 
 class prosmart_refmac(CPluginScript):
 
@@ -679,10 +682,10 @@ class prosmart_refmac(CPluginScript):
                 validateXMLPath = self.validate.makeFileName('PROGRAMXML')
                 validateXML = CCP4Utils.openFileToEtree(validateXMLPath)
                 if len(validateXML.xpath("//Validate_geometry_CCP4i2/Model_info"))>0:
-                   xml_validation.append(validateXML.xpath("//Validate_geometry_CCP4i2/Model_info")[0]) 
+                   xml_validation.append(validateXML.xpath("//Validate_geometry_CCP4i2/Model_info")[0])
                 if self.validate.container.controlParameters.DO_IRIS:
                    if len(validateXML.xpath("//Validate_geometry_CCP4i2/Iris"))>0:
-                      xml_validation.append(validateXML.xpath("//Validate_geometry_CCP4i2/Iris")[0]) 
+                      xml_validation.append(validateXML.xpath("//Validate_geometry_CCP4i2/Iris")[0])
                 if self.validate.container.controlParameters.DO_BFACT:
                    if len(validateXML.xpath("//Validate_geometry_CCP4i2/B_factors"))>0:
                       xml_validation.append(validateXML.xpath("//Validate_geometry_CCP4i2/B_factors")[0])
@@ -735,7 +738,6 @@ class prosmart_refmac(CPluginScript):
 
                        self.saveXml()
                    except:
-                       import traceback
                        print("Some problem with verdict...."); sys.stdout.flush()
                        exc_type, exc_value, exc_tb = sys.exc_info()[:3]
                        sys.stderr.write(str(exc_type) + '\n')
@@ -746,7 +748,6 @@ class prosmart_refmac(CPluginScript):
              except Exception as err:
                 xml_validation_status.text = "FAILURE"
                 self.saveXml()
-                import traceback
                 traceback.print_exc()
                 print("...Failed validation run after refinement", err)
 
@@ -755,6 +756,16 @@ class prosmart_refmac(CPluginScript):
             logfiles.append(self.firstRefmac.makeFileName('LOG'))
         if hasattr(self,"refmacPostCootPlugin"):
             logfiles.append(self.refmacPostCootPlugin.makeFileName('LOG'))
+
+        asuin = self.container.inputData.ASUIN
+        if asuin.isSet():
+            self.saveXml()
+            try:
+                xyzinPath = str(self.container.outputData.XYZOUT)
+                self.xmlroot.append(sequenceAlignment(xyzinPath, asuin))
+            except Exception as err:
+                traceback.print_exc()
+                print("...importing sequences for alignment test failed", err)
 
         self.createWarningsXML(logfiles)
         self.saveXml()