+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CCP4i2 Changelog
 
+## [2.5.0] - 2025-11-03
+
+- New task to check a model against AU contents
+- Standardised ModelCraft output file name
+- Twinning analysis in Servalcat report
+
 ## [2.4.3] - 2025-10-24
 
 - Improved checking/rebuilding of database and project-list XML files on startup